+++ src/Base/Http/Server.php
@@ -64,7 +64,7 @@ class Server
     */
     public function __get($name)
     {
-        return $this->{$name};
+        return $this->{$name} ?? null;
     }
 
 }